+++ src/borg/platform/__init__.py
@@ -6,6 +6,7 @@ Platform-specific APIs.
 Public APIs are documented in platform.base.
 """
 
+from .base import listxattr, getxattr, setxattr, ENOATTR
 from .base import acl_get, acl_set
 from .base import set_flags, get_flags
 from .base import SaveFile, SyncFile, sync_dir, fdatasync, safe_fadvise
@@ -21,12 +22,15 @@ if not sys.platform.startswith(('win32', )):
 
 if sys.platform.startswith('linux'):  # pragma: linux only
     from .linux import API_VERSION as OS_API_VERSION
+    from .linux import listxattr, getxattr, setxattr
     from .linux import acl_get, acl_set
     from .linux import set_flags, get_flags
     from .linux import SyncFile
 elif sys.platform.startswith('freebsd'):  # pragma: freebsd only
     from .freebsd import API_VERSION as OS_API_VERSION
+    from .freebsd import listxattr, getxattr, setxattr
     from .freebsd import acl_get, acl_set
 elif sys.platform == 'darwin':  # pragma: darwin only
     from .darwin import API_VERSION as OS_API_VERSION
+    from .darwin import listxattr, getxattr, setxattr
     from .darwin import acl_get, acl_set
