+++ src/borg/platform/freebsd.pyx
@@ -2,12 +2,28 @@ import os
 
 from ..helpers import posix_acl_use_stored_uid_gid
 from ..helpers import safe_encode, safe_decode
+from .xattr import _listxattr_inner, _getxattr_inner, _setxattr_inner, split_lstring
 
-API_VERSION = '1.1_03'
+API_VERSION = '1.2_01'
 
 cdef extern from "errno.h":
     int errno
     int EINVAL
+
+cdef extern from "sys/extattr.h":
+    ssize_t c_extattr_list_file "extattr_list_file" (const char *path, int attrnamespace, void *data, size_t nbytes)
+    ssize_t c_extattr_list_link "extattr_list_link" (const char *path, int attrnamespace, void *data, size_t nbytes)
+    ssize_t c_extattr_list_fd "extattr_list_fd" (int fd, int attrnamespace, void *data, size_t nbytes)
+
+    ssize_t c_extattr_get_file "extattr_get_file" (const char *path, int attrnamespace, const char *attrname, void *data, size_t nbytes)
+    ssize_t c_extattr_get_link "extattr_get_link" (const char *path, int attrnamespace, const char *attrname, void *data, size_t nbytes)
+    ssize_t c_extattr_get_fd "extattr_get_fd" (int fd, int attrnamespace, const char *attrname, void *data, size_t nbytes)
+
+    int c_extattr_set_file "extattr_set_file" (const char *path, int attrnamespace, const char *attrname, const void *data, size_t nbytes)
+    int c_extattr_set_link "extattr_set_link" (const char *path, int attrnamespace, const char *attrname, const void *data, size_t nbytes)
+    int c_extattr_set_fd "extattr_set_fd" (int fd, int attrnamespace, const char *attrname, const void *data, size_t nbytes)
+
+    int EXTATTR_NAMESPACE_USER
 
 cdef extern from "sys/types.h":
     int ACL_TYPE_ACCESS
@@ -30,6 +46,50 @@ cdef extern from "sys/acl.h":
 cdef extern from "unistd.h":
     long lpathconf(const char *path, int name)
     int _PC_ACL_NFS4
+
+
+def listxattr(path, *, follow_symlinks=True):
+    def func(path, buf, size):
+        cdef char *buffer = <char *> buf
+        if isinstance(path, int):
+            return c_extattr_list_fd(path, EXTATTR_NAMESPACE_USER, buffer, size)
+        else:
+            if follow_symlinks:
+                return c_extattr_list_file(path, EXTATTR_NAMESPACE_USER, buffer, size)
+            else:
+                return c_extattr_list_link(path, EXTATTR_NAMESPACE_USER, buffer, size)
+
+    n, buf = _listxattr_inner(func, path)
+    return [os.fsdecode(name) for name in split_lstring(buf[:n]) if name]
+
+
+def getxattr(path, name, *, follow_symlinks=True):
+    def func(path, name, buf, size):
+        cdef char *buffer = <char *> buf
+        if isinstance(path, int):
+            return c_extattr_get_fd(path, EXTATTR_NAMESPACE_USER, name, buffer, size)
+        else:
+            if follow_symlinks:
+                return c_extattr_get_file(path, EXTATTR_NAMESPACE_USER, name, buffer, size)
+            else:
+                return c_extattr_get_link(path, EXTATTR_NAMESPACE_USER, name, buffer, size)
+
+    n, buf = _getxattr_inner(func, path, name)
+    return buf[:n] or None
+
+
+def setxattr(path, name, value, *, follow_symlinks=True):
+    def func(path, name, value, size):
+        cdef char *val = NULL if value is None else <char *> value
+        if isinstance(path, int):
+            return c_extattr_set_fd(path, EXTATTR_NAMESPACE_USER, name, val, size)
+        else:
+            if follow_symlinks:
+                return c_extattr_set_file(path, EXTATTR_NAMESPACE_USER, name, val, size)
+            else:
+                return c_extattr_set_link(path, EXTATTR_NAMESPACE_USER, name, val, size)
+
+    _setxattr_inner(func, path, name, value)
 
 
 cdef _get_acl(p, type, item, attribute, int flags):
